flask/requirements.txt
Flask
requests
pymongo
flask_pymongo
pandas
unidecode
openai
python-dotenv